relocate client files and redirect to homepage after signup
