+++ api/app/controllers/api/v1/profiles_controller.rb
@@ -1,0 +1,9 @@
+module Api
+  module V1
+    class ProfilesController < Api::V1::BaseController
+      def show
+        respond_with current_resource_owner
+      end
+    end
+  end
+end
