+++ api/config/routes.rb
@@ -5,6 +5,7 @@ Rails.application.routes.draw do
   namespace :api do
     namespace :v1, defaults: { format: :json } do
       resource :health, only: [:show]
+      resource :profile, only: [:show]
     end
   end
 end
